+++ app/assets/stylesheets/devise/modal.sass
@@ -1,0 +1,37 @@
+@import url('https://fonts.googleapis.com/css2?family=Merriweather:wght@700&family=Montserrat&display=swap');
+
+*
+  font-family: 'Merriweather', serif;
+  font-family: 'Montserrat', sans-serif;
+
+.modal-backdrop
+  opacity: 0.1 !important
+
+.modal-content
+  width: 420px !important
+
+#close-modal-cross
+  font-size: 17px
+  color: #005164
+
+.modalInput
+  border-radius: 2px
+
+.text-muted
+  font-size: 17px
+  color: #7F8374
+
+#modalSubmitButton
+  background-color: #3A5637
+  border-radius: 2px
+
+.modal-title
+  font: 24px
+  color: #7F8374
+
+.footer-link
+  color: #005164
+  font: 14px
+
+#navLink
+  color: #005164
