+++ app/controllers/application_controller.rb
@@ -1,19 +1,9 @@
 class ApplicationController < ActionController::Base
-  helper_method :resource_name, :resource, :devise_mapping, :resource_class
+  before_action :configure_permitted_parameters, if: :devise_controller?
 
-  def resource_name
-    :user
-  end
+  protected
 
-  def resource
-    @resource ||= User.new
-  end
-
-  def resource_class
-    User
-  end
-
-  def devise_mapping
-    @devise_mapping ||= Devise.mappings[:user]
+  def configure_permitted_parameters
+    devise_parameter_sanitizer.permit(:sign_up, keys: [:first_name, :last_name, :phone_number])
   end
 end
