+++ app/views/devise/sessions/create.js.erb
@@ -1,0 +1,4 @@
+const modal = document.querySelector('.modal')
+$(document).ready(function() {
+  $('.modal').modal('hide');
+})
