+++ app/views/devise/sessions/new.html.erb
@@ -1,5 +1,3 @@
-<h2>Log in</h2>
-
 <%= form_for(resource, as: resource_name, url: session_path(resource_name)) do |f| %>
   <div class="field">
     <%= f.label :email %><br />
