+++ app/views/devise/sessions/new.js.erb
@@ -1,0 +1,2 @@
+const errorMessage = document.querySelector('.errors')
+errorMessage.innerHTML = "Incorrect Email/Password. \<br>Please review the errors below:"
