--- app/views/shared/_login_form.html.erb
@@ -1,43 +1,0 @@
-<div class="modal fade" id="loginModal" tabindex="-1" role="dialog" aria-labellebdy="loginModal" aria-hidden="true">
-  <div class="modal-dialog modal-dialog-centered" role="document">
-    <div class="modal-content">
-
-      <div class="modal-header">
-        <h5 class="modal-title" id="loginModal">LOG IN</h5>
-        <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-          <span aria-hidden="true">x</span>
-        </button>
-      </div>
-
-      <div class="modal-body">
-        <div class="container-fluid">
-          <%= form_for(:user, :url => session_path(:user)) do |f| %>
-
-            <div class="form-group">
-              <small class="form-text text-muted">Email</small>
-              <%= f.text_field :email, class: "form-control" %>
-            </div>
-
-            <div class="form-group">
-              <small  class="form-text text-muted">Password</small>
-              <%= f.password_field :password, class: "form-control"  %>
-            </div>
-
-            <% if devise_mapping.rememberable? %>
-              <div class="form-group form-check">
-                <input type="checkbox" class="form-check-input" id="loginCheck">
-                <label class="form-check-label" for="loginCheck">Remember me</label>
-              </div>
-            <% end %>
-
-            <div class="actions">
-              <%= f.submit "Log in", class: "btn btn-primary mt-2"  %>
-            </div>
-          <% end %>
-          <%= link_to "or Sign up?", new_user_registration_path %> <br>
-          <%= link_to "Forgot your password?", new_password_path(:user) %>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
