+++ config/routes.rb
@@ -1,6 +1,8 @@
 Rails.application.routes.draw do
   get 'welcome/home'
-  devise_for :users
+  devise_for :users, controllers: {
+          sessions: 'users/sessions'
+  }
 
   root 'welcome#home', as: 'home'
